chore: remove duplications and clean code (#269)
* chore: remove unused dashboard translation keys

* chore: extract inline components and fix inverted schema dependency

- Extract SaveTemplateDrawer, LocationMapPreview, LocationCardSkeleton,
  AccountNavLink into their own files
- Move zFormFieldsOrg/FormFieldsOrg from page-organization-new to schema.ts
- Fix form-organization importing FormFieldsOrg from a page file
- Move locationsInfiniteOptions to location-queries.ts (was in page file,
  imported by route)



* chore: deduplicate and relocate shared Zod schemas (#270)

Move zUserSummary to user/schema and zLocationSummary to location/schema,
deduplicate inline enums in zStopPassenger by reusing zRequestStatus/zTripType,
and remove circular dependency between commute and booking schemas.
com>

---------
