+++ src/locales/en/dashboard.json
@@ -1,11 +1,8 @@
 {
   "title": "Dashboard",
   "newCommuteAction": "New Commute",
-  "requestCommuteAction": "Request a Commute",
-  "noMatchingRide": "No ride suits you?",
   "today": "Today",
   "noCommutesForDay": "No commutes for this day",
-  "noCommutesForWeek": "No commutes this week",
   "booking": {
     "title": "Book a ride",
     "selectStop": "Select a stop",
