+++ src/locales/fr/dashboard.json
@@ -1,11 +1,8 @@
 {
   "title": "Tableau de bord",
   "newCommuteAction": "Nouveau trajet",
-  "requestCommuteAction": "Demander un trajet",
-  "noMatchingRide": "Aucun trajet ne vous convient ?",
   "today": "Aujourd'hui",
   "noCommutesForDay": "Aucun trajet pour ce jour",
-  "noCommutesForWeek": "Aucun trajet cette semaine",
   "booking": {
     "title": "Réserver un trajet",
     "selectStop": "Sélectionner un arrêt",
